moved the tag reguest path creation methods out of the LogixDriver class to functions in the request packets module removed forward open requirement for the get_module_info method, it's not needed
